build-debian: Use manual docker build for source package
This allows us to have more control on the volumes we mount in
docker and on the location of the source path.

And permits us to properly mount the parent folder of the source
that we need to access to read the .orig files for non-native
packages.

It comes as a little extra cost, but also it give us way more
control on the builder instance.